Improved keyboard input and mount SD card as drive C:

+++ src/fake86/fake86.h
@@ -11,6 +11,6 @@ void getactivepalette(uint8_t** palette, int* paletteSize);
 
 void handlekeydown(uint16_t scancode);
 void handlekeyup(uint16_t scancode);
-void handlekeydownraw(uint8_t scancode);
-void handlekeyupraw(uint8_t scancode);
+void handlekeydownraw(uint16_t scancode);
+void handlekeyupraw(uint16_t scancode);
